requirements.txt
python-dotenv
langchain
langgraph
langchain-openai
pandas
streamlit
langgraph-checkpoint-sqlite  
langchain
langchain-pinecone
pinecone-client